# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentGuard Core Engine",
    "description": "Framework-Agnostic Runtime Governance Layer for Autonomous AI Agents",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/guard/check": {
      "post": {
        "tags": [
          "guard"
        ],
        "summary": "Guard Check",
        "description": "Evaluate a proposed tool call through the PolicyEngine chokepoint.\n\nStrict execution flow:\n  1. Authenticate agent & construct CallerIdentity\n  2. Call deterministic PolicyEngine (via Phase 6 Redis-wired factory)\n  3. Create ToolRequest, AuditLog (with monotonic sequence_number), and HITLRequest if PENDING\n  4. db.commit() -- HARD GATE: if commit fails, abort with 500 (mock tool never called)\n  5. If ALLOW and mock handler exists, execute safe mock tool\n  6. Return decision response to agent",
        "operationId": "guard_check_api_v1_guard_check_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dashboard/stats": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Dashboard Stats",
        "description": "Return aggregate numbers for the dashboard home screen.\n\nAll values are scoped to the caller's organization. A single pass over\ntool_requests is used for all decision-count metrics to avoid N+1.",
        "operationId": "get_dashboard_stats_api_v1_dashboard_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/dashboard/activity": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Get Dashboard Activity",
        "description": "Return a reverse-chronological feed of tool requests for polling.\n\nUses a single LEFT JOIN query with count to avoid N+1. The frontend should\npoll this endpoint; WebSockets/SSE are explicitly deferred to Phase 14.",
        "operationId": "get_dashboard_activity_api_v1_dashboard_activity_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset (cursor)",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset (cursor)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityFeed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityFeed": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActivityItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "ActivityFeed"
      },
      "ActivityItem": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Name"
          },
          "tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name"
          },
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "agent_id",
          "decision",
          "created_at"
        ],
        "title": "ActivityItem",
        "description": "Single item in the reverse-chronological activity feed."
      },
...
      "DashboardStats": {
        "properties": {
          "total_agents": {
            "type": "integer",
            "title": "Total Agents"
          },
          "total_tools": {
            "type": "integer",
            "title": "Total Tools"
          },
          "requests_today": {
            "type": "integer",
            "title": "Requests Today"
          },
          "allowed_today": {
            "type": "integer",
            "title": "Allowed Today"
          },
          "blocked_today": {
            "type": "integer",
            "title": "Blocked Today"
          },
          "pending_today": {
            "type": "integer",
            "title": "Pending Today"
          },
          "total_spend_today": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Spend Today"
          },
          "budget_utilization": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Budget Utilization"
          }
        },
        "type": "object",
        "required": [
          "total_agents",
          "total_tools",
          "requests_today",
          "allowed_today",
          "blocked_today",
          "pending_today",
          "total_spend_today",
          "budget_utilization"
        ],
        "title": "DashboardStats",
        "description": "Aggregate numbers for the dashboard home screen, org-scoped."
      },
      "DecisionEnum": {
        "type": "string",
        "enum": [
          "ALLOW",
          "DENY",
          "PENDING"
        ],
        "title": "DecisionEnum"
      },
      "DecisionInput": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "tool_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Tool Name"
          },
          "action": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Action"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "estimated_tokens": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Estimated Tokens",
            "default": 0
          },
          "estimated_cost": {
            "type": "number",
            "minimum": 0.0,
            "title": "Estimated Cost",
            "default": 0.0
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "tool_name",
          "action"
        ],
        "title": "DecisionInput",
        "description": "Input payload to the Policy Engine matching project.md contract."
      },
      "GuardResponse": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/DecisionEnum"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "request_id",
          "reason"
        ],
        "title": "GuardResponse",
        "description": "HTTP response shape for POST /guard/check, matching project.md contract."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}